Changes to be committed: 	modified:   src/main/java/com/cassess/CassessApplication.java 	new file:   src/main/java/com/cassess/config/AppConfig.java 	new file:   src/main/java/com/cassess/config/DBConfig.java 	new file:   src/main/java/com/cassess/config/WebAppInitializer.java 	new file:   src/main/java/com/cassess/controller/AppController.java 	renamed:    src/main/java/com/cassess/model/CAssessDAO.java -> src/main/java/com/cassess/dao/CAssessDAO.java 	renamed:    src/main/java/com/cassess/model/slack/ConsumeChannels.java -> src/main/java/com/cassess/dao/slack/ConsumeChannels.java 	renamed:    src/main/java/com/cassess/model/slack/ConsumeGroups.java -> src/main/java/com/cassess/dao/slack/ConsumeGroups.java 	new file:   src/main/java/com/cassess/dao/slack/ConsumeTeam.java 	renamed:    src/main/java/com/cassess/model/slack/ConsumeUsers.java -> src/main/java/com/cassess/dao/slack/ConsumeUsers.java 	new file:   src/main/java/com/cassess/dao/slack/ISlackDAO.java 	new file:   src/main/java/com/cassess/dao/slack/SlackDAO.java 	renamed:    src/main/java/com/cassess/model/slack/ChannelObject.java -> src/main/java/com/cassess/entity/slack/ChannelObject.java 	renamed:    src/main/java/com/cassess/model/slack/ChannelsList.java -> src/main/java/com/cassess/entity/slack/ChannelsList.java 	renamed:    src/main/java/com/cassess/model/slack/GroupList.java -> src/main/java/com/cassess/entity/slack/GroupList.java 	renamed:    src/main/java/com/cassess/model/slack/GroupObject.java -> src/main/java/com/cassess/entity/slack/GroupObject.java 	renamed:    src/main/java/com/cassess/model/slack/SlackAuth.java -> src/main/java/com/cassess/entity/slack/SlackAuth.java 	renamed:    src/main/java/com/cassess/model/slack/SlackGroupObject.java -> src/main/java/com/cassess/entity/slack/SlackGroupObject.java 	renamed:    src/main/java/com/cassess/model/slack/SlackIcon.java -> src/main/java/com/cassess/entity/slack/SlackIcon.java 	renamed:    src/main/java/com/cassess/model/slack/SlackMessage.java -> src/main/java/com/cassess/entity/slack/SlackMessage.java 	renamed:    src/main/java/com/cassess/model/slack/TeamInfo.java -> src/main/java/com/cassess/entity/slack/TeamInfo.java 	renamed:    src/main/java/com/cassess/model/slack/TeamObject.java -> src/main/java/com/cassess/entity/slack/TeamObject.java 	renamed:    src/main/java/com/cassess/model/slack/UserInfo.java -> src/main/java/com/cassess/entity/slack/UserInfo.java 	renamed:    src/main/java/com/cassess/model/slack/UserList.java -> src/main/java/com/cassess/entity/slack/UserList.java 	renamed:    src/main/java/com/cassess/model/slack/UserObject.java -> src/main/java/com/cassess/entity/slack/UserObject.java 	renamed:    src/main/java/com/cassess/model/slack/UserProfile.java -> src/main/java/com/cassess/entity/slack/UserProfile.java 	deleted:    src/main/java/com/cassess/model/slack/ConsumeTeam.java 	modified:   src/main/java/com/cassess/model/taiga/ConsumeAuthUser.java 	modified:   src/main/java/com/cassess/model/taiga/TaskCount.java 	modified:   src/main/java/com/cassess/service/ApiServiceImpl.java 	modified:   src/main/java/com/cassess/service/DAO/SlackServiceDAO.java 	modified:   src/main/java/com/cassess/service/DAO/SlackServiceDAOImpl.java 	new file:   src/main/java/com/cassess/service/ISlackService.java 	new file:   src/main/java/com/cassess/service/SlackService.java 	modified:   src/main/resources/application.properties 	deleted:    src/main/resources/applicationContext.xml 	modified:   src/main/resources/static/js/hello.js 	new file:   src/main/resources/static/partials/slack-user.html 	modified:   src/main/webapp/WEB-INF/dispatcher-servlet.xml
