+++ src/main/java/com/cassess/CassessApplication.java
@@ -1,5 +1,7 @@
 package com.cassess;
 
+import com.cassess.dao.slack.ConsumeChannels;
+import com.cassess.dao.slack.ConsumeUsers;
 import com.cassess.model.github.GatherGitHubData;
 import com.cassess.model.taiga.ConsumeAuthUser;
 import com.cassess.model.taiga.ConsumeProjectList;
@@ -13,8 +15,6 @@ import org.springframework.boot.autoconfigure.security.SecurityProperties;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.ImportResource;
-import org.springframework.context.support.ClassPathXmlApplicationContext;
 import org.springframework.core.annotation.Order;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
@@ -23,9 +23,6 @@ import org.springframework.security.web.csrf.CookieCsrfTokenRepository;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.cassess.model.slack.ConsumeChannels;
-import com.cassess.model.slack.ConsumeUsers;
-
 import java.security.Principal;
 import java.util.HashMap;
 import java.util.Map;
@@ -33,7 +30,6 @@ import java.util.UUID;
 
 @SpringBootApplication
 @RestController
-@ImportResource({"classpath*:applicationContext.xml"})
 public class CassessApplication {
 
    /* @RequestMapping("/resource")
@@ -71,19 +67,19 @@ public class CassessApplication {
     public CommandLineRunner commandLineRunner(ApplicationContext ctx) {
         return args -> {
             System.out.println("Let's inspect the beans provided by Spring Boot:");
-
+            /*
     		ConsumeChannels consumeChannels = (ConsumeChannels) ctx.getBean("consumeChannels");
     		consumeChannels.getChannelsList();
 
             ConsumeUsers consumeUsers = (ConsumeUsers) ctx.getBean("consumeUsers");
     		consumeUsers.getUserInfo("U2G79FELT");
-
+    		
             ConsumeAuthUser consumeAuthUser = (ConsumeAuthUser) ctx.getBean("consumeAuthUser");
             AuthUser auth = consumeAuthUser.getUserInfo();
-
+        	
             ConsumeProjectList consumeProjectList = (ConsumeProjectList) ctx.getBean("consumeProjectList");
             Project proj = consumeProjectList.getProjectInfo(auth.getAuth_token(), auth.getId());
-
+            
             GetTaskData getTaskData = (GetTaskData) ctx.getBean("getTaskData");
             getTaskData.getTasks(proj.getId(), auth.getAuth_token(), 1);
             getTaskData.getMembers(proj.getId(), auth.getAuth_token(), 1);
@@ -91,12 +87,11 @@ public class CassessApplication {
             GatherGitHubData gatherGitHubData = (GatherGitHubData) ctx.getBean("gatherGitHubData");
             gatherGitHubData.fetchData();
             //get commit List returns all commits there are
-            System.out.println(gatherGitHubData.getCommitList());
+            System.out.println(gatherGitHubData.getCommitList());*/
         };
     }
 
     public static void main(String[] args) {
-        //ClassPathXmlApplicationContext context = new ClassPathXmlApplicationContext("classpath:applicationContext.xml");
         SpringApplication.run(CassessApplication.class, args);
     }
