+++ src/main/java/com/cassess/config/AppConfig.java
@@ -1,0 +1,34 @@
+package com.cassess.config;
+/*
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.DefaultServletHandlerConfigurer;
+import org.springframework.web.servlet.config.annotation.EnableWebMvc;
+import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
+import org.springframework.web.servlet.view.InternalResourceViewResolver;
+
+@Configuration 
+@ComponentScan("com") 
+@EnableWebMvc   
+public class AppConfig extends WebMvcConfigurerAdapter  { 
+	
+    @Bean  
+    public InternalResourceViewResolver viewResolver() {  
+	InternalResourceViewResolver resolver = new InternalResourceViewResolver();  
+        resolver.setPrefix("/WEB-INF/");  
+        resolver.setSuffix(".html");
+        return resolver;  
+    }
+    
+    @Override
+    public void configureDefaultServletHandling(DefaultServletHandlerConfigurer config) {
+    	config.enable();
+    }
+    
+    @Override
+    public void addResourceHandlers(ResourceHandlerRegistry registry) {
+    	registry.addResourceHandler("/app-resources/**").addResourceLocations("/resources/");
+    }    
+} */
