+++ src/main/java/com/cassess/config/WebAppInitializer.java
@@ -1,0 +1,18 @@
+package com.cassess.config;
+/*
+import org.springframework.web.servlet.support.AbstractAnnotationConfigDispatcherServletInitializer;
+
+public class WebAppInitializer extends AbstractAnnotationConfigDispatcherServletInitializer {
+	@Override
+	protected Class<?>[] getRootConfigClasses() {
+		return new Class[] { AppConfig.class };
+	}
+	@Override
+	protected Class<?>[] getServletConfigClasses() {
+		return null;
+	}
+	@Override
+	protected String[] getServletMappings() {
+		return new String[] { "/" };
+	}
+} */
