+++ src/main/java/com/cassess/dao/slack/ConsumeChannels.java
@@ -1,11 +1,14 @@
-package com.cassess.model.slack;
+package com.cassess.dao.slack;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.client.RestTemplate;
 
-import com.cassess.model.CAssessDAO;
+import com.cassess.dao.CAssessDAO;
+import com.cassess.entity.slack.ChannelObject;
+import com.cassess.entity.slack.ChannelsList;
+import com.cassess.entity.slack.SlackAuth;
 
 @Service
 @Transactional
