+++ src/main/java/com/cassess/dao/slack/ConsumeTeam.java
@@ -1,0 +1,5 @@
+package com.cassess.dao.slack;
+
+public class ConsumeTeam {
+
+}
