+++ src/main/java/com/cassess/dao/slack/ConsumeUsers.java
@@ -1,4 +1,4 @@
-package com.cassess.model.slack;
+package com.cassess.dao.slack;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -6,7 +6,11 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.client.RestTemplate;
 
-import com.cassess.model.CAssessDAO;
+import com.cassess.dao.CAssessDAO;
+import com.cassess.entity.slack.SlackAuth;
+import com.cassess.entity.slack.UserInfo;
+import com.cassess.entity.slack.UserList;
+import com.cassess.entity.slack.UserObject;
 
 @Service
 @Transactional
