--- src/main/java/com/cassess/model/slack/ConsumeTeam.java
@@ -1,5 +1,0 @@
-package com.cassess.model.slack;
-
-public class ConsumeTeam {
-
-}
