--- src/main/resources/applicationContext.xml
@@ -1,93 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<beans xmlns="http://www.springframework.org/schema/beans"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" 
-	xmlns:context="http://www.springframework.org/schema/context"
-	xmlns:tx="http://www.springframework.org/schema/tx" 
-	xmlns:mvc="http://www.springframework.org/schema/mvc"
-	xsi:schemaLocation="http://www.springframework.org/schema/beans 
-                           http://www.springframework.org/schema/beans/spring-beans.xsd
-                           http://www.springframework.org/schema/tx 
-                           http://www.springframework.org/schema/tx/spring-tx.xsd
-                           http://www.springframework.org/schema/context 
-                           http://www.springframework.org/schema/context/spring-context.xsd
-                        http://www.springframework.org/schema/mvc
-						http://www.springframework.org/schema/mvc/spring-mvc.xsd">
-
-	<context:annotation-config />
-	<context:component-scan base-package="com" />
-	
-	    <!-- Add support for conversion, formatting and validation support -->
-    <mvc:annotation-driven/>
-
-    <!-- Define Spring MVC view resolver -->
-    <bean class="org.springframework.web.servlet.view.InternalResourceViewResolver">
-        <property name="prefix" value="/WEB-INF/view/" />
-        <property name="suffix" value=".jsp" />
-    </bean>
-	
-	<!-- 
-	<context:property-placeholder
-		ignore-unresolvable="true" location="classpath:config.properties" />
-	-->
-		
-	<bean id="dataSource" class="org.apache.commons.dbcp.BasicDataSource" destroy-method="close">
-		<property name="driverClassName" value="com.mysql.jdbc.Driver" />
-		<property name="url" value="jdbc:mysql://localhost/cassess" />
-		<property name="username" value="root" />
-		<property name="password" value="root123" />
-	</bean>
-	
-	<!-- 
-	<bean id="mySessionFactory" class="org.springframework.orm.hibernate5.LocalSessionFactoryBean">
-		<property name="dataSource" ref="dataSource" />
-		<property name="mappingResources">
-			<list>
-			</list>
-		</property>
-		<property name="hibernateProperties">
-			<value>
-				hibernate.dialect=org.hibernate.dialect.HSQLDialect
-			</value>
-		</property>
-	</bean>
-	-->
-	
-	<bean id="emf" class="org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean">
-		<property name="dataSource" ref="dataSource" />
-		<!-- <property name="persistenceUnitName" value="default" /> -->
-		<property name="packagesToScan" value="com.cassess.*" />
-		<property name="jpaVendorAdapter">
-			<bean class="org.springframework.orm.jpa.vendor.HibernateJpaVendorAdapter" />
-		</property>
-		<property name="jpaProperties">
-			<props>
-				<prop key="hibernate.hbm2ddl.auto">create</prop>
-
-				<prop key="hibernate.dialect">org.hibernate.dialect.MySQLDialect</prop>
-			</props>
-		</property>
-	</bean>
-	
-	<bean id="transactionManager" class="org.springframework.orm.jpa.JpaTransactionManager">
-		<property name="entityManagerFactory" ref="emf" />
-	</bean>
-	
-	<bean id="metadataUtil"
-		class="com.googlecode.genericdao.search.jpa.JPAAnnotationMetadataUtil">
-	</bean>
-	
-	<bean id="searchProcessor" class="com.googlecode.genericdao.search.jpa.JPASearchProcessor">
-		<constructor-arg ref="metadataUtil" />
-	</bean>
-	
-	<bean id="consumeUsers" class="com.cassess.model.slack.ConsumeUsers">
-	</bean>
-
-	<bean id="consumeProjectList" class="com.cassess.model.taiga.ConsumeProjectList">
-	</bean>
-
-    <bean id="consumeAuthUser" class="com.cassess.model.taiga.ConsumeAuthUser">
-    </bean>
-	
-	<tx:annotation-driven transaction-manager="transactionManager" />
-</beans>
