+++ src/main/resources/static/partials/slack-user.html
@@ -1,0 +1,10 @@
+<div ng-controller="slack" ng-show="authenticated">
+<table class="table table-striped">
+<tr><th>ID </th><th> Real Name </th><th> Time Zone</th></tr>
+<tr ng-repeat="row in slackData">
+<td><span ng-bind="row.id"></span></td>
+<td><span ng-bind="row.real_name"></span></td>
+<td><span ng-bind="row.tz_label"></span></td>
+</tr>
+</table>
+</div>
