+++ src/main/webapp/WEB-INF/dispatcher-servlet.xml
@@ -12,61 +12,8 @@
 		http://www.springframework.org/schema/mvc
 		http://www.springframework.org/schema/mvc/spring-mvc.xsd
 		http://www.springframework.org/schema/tx
-		http://www.springframework.org/schema/tx/spring-tx.xsd">
-
-    <!--&lt;!&ndash; Add support for component scanning &ndash;&gt;-->
-    <!--<context:component-scan base-package="com.cassess" />-->
-
-    <!--&lt;!&ndash; Add support for conversion, formatting and validation support &ndash;&gt;-->
-    <!--<mvc:annotation-driven/>-->
-
-    <!--&lt;!&ndash; Define Spring MVC view resolver &ndash;&gt;-->
-    <!--<bean-->
-            <!--class="org.springframework.web.servlet.view.InternalResourceViewResolver">-->
-        <!--<property name="prefix" value="/WEB-INF/view/" />-->
-        <!--<property name="suffix" value=".jsp" />-->
-    <!--</bean>-->
-
-    <!--&lt;!&ndash; Define Database DataSource / connection pool &ndash;&gt;-->
-    <!--<bean id="myDataSource" class="com.mchange.v2.c3p0.ComboPooledDataSource"-->
-          <!--destroy-method="close">-->
-        <!--<property name="driverClass" value="com.mysql.jdbc.Driver" />-->
-        <!--<property name="jdbcUrl" value="jdbc:mysql://localhost:3306/cassess?useSSL=false" />-->
-        <!--<property name="user" value="TestAccount" />-->
-        <!--<property name="password" value="TestAccount!23" />-->
-
-        <!--&lt;!&ndash; these are connection pool properties for C3P0 &ndash;&gt;-->
-        <!--<property name="minPoolSize" value="5" />-->
-        <!--<property name="maxPoolSize" value="20" />-->
-        <!--<property name="maxIdleTime" value="30000" />-->
-    <!--</bean>-->
-
-    <!--&lt;!&ndash; Setup Hibernate session factory &ndash;&gt;-->
-    <!--<bean id="sessionFactory"-->
-          <!--class="org.springframework.orm.hibernate5.LocalSessionFactoryBean">-->
-        <!--<property name="dataSource" ref="myDataSource" />-->
-        <!--<property name="packagesToScan" value="com.cassess" />-->
-        <!--<property name="hibernateProperties">-->
-            <!--<props>-->
-                <!--<prop key="hibernate.dialect">org.hibernate.dialect.MySQLDialect</prop>-->
-                <!--<prop key="hibernate.show_sql">true</prop>-->
-            <!--</props>-->
-        <!--</property>-->
-    <!--</bean>-->
-
-    <!--&lt;!&ndash; Setup Hibernate transaction manager &ndash;&gt;-->
-    <!--<bean id="myTransactionManager"-->
-          <!--class="org.springframework.orm.hibernate5.HibernateTransactionManager">-->
-        <!--<property name="sessionFactory" ref="sessionFactory"/>-->
-    <!--</bean>-->
-   <!---->
-
-    <!--&lt;!&ndash; Enable configuration of transactional behavior based on annotations &ndash;&gt;-->
-    <!--<tx:annotation-driven transaction-manager="myTransactionManager" />
-    -->
-
-</beans>
-
+		http://www.springframework.org/schema/tx/spring-tx.xsd"
+/>
